requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
Django==3.2.23
django-filter==23.5
django-jalali==6.0.1
djangorestframework==3.14.0
h11==0.14.0
idna==3.6
jdatetime==4.1.1
Markdown==3.5.2
outcome==1.3.0.post0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
pytz==2023.3.post1
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
